training/requirements.txt
torch>=2.0.1
opencv-python>=4.8.0.76
Pillow>=9.4.0
numpy>=1.23.5
PyYAML>=6.0.1
onnx>=1.14.1
tqdm>=4.66.1
Cython>=3.0.2